+++ src/platform/native.rs
@@ -29,6 +29,12 @@ use crate::{Context, ContextBuilder, State};
 
 pub type GlContext = glow::native::Context;
 
+pub const DEFAULT_VERTEX_SHADER: &str =
+    concat!("#version 150\n", include_str!("../resources/shader.vert"));
+
+pub const DEFAULT_FRAGMENT_SHADER: &str =
+    concat!("#version 150\n", include_str!("../resources/shader.frag"));
+
 pub struct Platform {
     sdl: Sdl,
     window: Window,