+++ src/platform/web.rs
@@ -14,6 +14,16 @@ use crate::{Context, ContextBuilder, State};
 
 pub type GlContext = glow::web::Context;
 
+pub const DEFAULT_VERTEX_SHADER: &str = concat!(
+    "#version 300 es\n",
+    include_str!("../resources/shader.vert")
+);
+
+pub const DEFAULT_FRAGMENT_SHADER: &str = concat!(
+    "#version 300 es\nprecision mediump float;\n",
+    include_str!("../resources/shader.frag")
+);
+
 pub struct Platform {}
 
 impl Platform {